+++ app/code/Magento/InventoryCatalog/Test/Integration/InventoryReservationsApi/ApplyDataToLegacyStockItemAtReservationPlacingTest.php
@@ -9,7 +9,7 @@ namespace Magento\InventoryCatalog\Test\Integration\InventoryReservationsApi;
 
 use Magento\Catalog\Api\ProductRepositoryInterface;
 use Magento\InventoryReservations\Model\CleanupReservationsInterface;
-use Magento\InventoryReservationsApi\Api\ReservationBuilderInterface;
+use Magento\InventoryReservations\Model\ReservationBuilderInterface;
 use Magento\InventoryReservationsApi\Api\AppendReservationsInterface;
 use Magento\InventoryCatalog\Api\DefaultStockProviderInterface;
 use PHPUnit\Framework\TestCase;
