+++ app/code/Magento/InventoryReservations/etc/di.xml
@@ -6,15 +6,14 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <!-- Reservation -->
-    <preference for="Magento\InventoryReservationsApi\Api\AppendReservationsInterface" type="Magento\InventoryReservations\Model\Reservation\Command\AppendReservations"/>
-    <preference for="Magento\InventoryReservationsApi\Api\ReservationBuilderInterface" type="Magento\InventoryReservations\Model\Reservation\ReservationBuilder"/>
+    <preference for="Magento\InventoryReservationsApi\Api\AppendReservationsInterface" type="Magento\InventoryReservations\Model\AppendReservations"/>
     <preference for="Magento\InventoryReservationsApi\Api\Data\ReservationInterface" type="Magento\InventoryReservations\Model\Reservation"/>
-    <preference for="Magento\InventoryReservations\Model\CleanupReservationsInterface" type="Magento\InventoryReservations\Model\ResourceModel\Reservation\CleanupReservations"/>
-    <type name="Magento\InventoryReservations\Model\ResourceModel\Reservation\CleanupReservations">
+    <preference for="Magento\InventoryReservations\Model\ReservationBuilderInterface" type="Magento\InventoryReservations\Model\ReservationBuilder"/>
+    <preference for="Magento\InventoryReservations\Model\CleanupReservationsInterface" type="Magento\InventoryReservations\Model\ResourceModel\CleanupReservations"/>
+    <type name="Magento\InventoryReservations\Model\ResourceModel\CleanupReservations">
         <arguments>
             <argument name="groupConcatMaxLen" xsi:type="number">2000</argument>
         </arguments>
     </type>
-    <preference for="Magento\InventoryReservations\Model\GetReservationsQuantityInterface" type="Magento\InventoryReservations\Model\ResourceModel\Reservation\GetReservationsQuantity"/>
+    <preference for="Magento\InventoryReservations\Model\GetReservationsQuantityInterface" type="Magento\InventoryReservations\Model\ResourceModel\GetReservationsQuantity"/>
 </config>
