+++ app/code/Magento/InventorySales/Test/Integration/Stock/GetProductSalableQtyTest.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
 namespace Magento\InventorySales\Test\Integration\Stock;
 
 use Magento\InventoryReservations\Model\CleanupReservationsInterface;
-use Magento\InventoryReservationsApi\Api\ReservationBuilderInterface;
+use Magento\InventoryReservations\Model\ReservationBuilderInterface;
 use Magento\InventoryReservationsApi\Api\AppendReservationsInterface;
 use Magento\InventorySalesApi\Api\GetProductSalableQtyInterface;
 use Magento\TestFramework\Helper\Bootstrap;
